+++ src/main/java/com/example/music_player/activeMQ/IListenerService.java
@@ -1,10 +1,7 @@
 package com.example.music_player.activeMQ;
 
-
 import com.example.music_player.entity.Source;
 
-import java.io.IOException;
-
 public interface IListenerService {
-
+    void getFileBySourceFromListener(Source sourceFromListener);
 }
